+++ student-distrib/system_calls.h
@@ -3,7 +3,7 @@
 
 #include "types.h"
 
-extern int32_t sys_halt(uint8_t status);
+extern int32_t sys_halt(uint32_t status);
 extern int32_t sys_execute(const uint8_t* command);
 extern int32_t sys_read(int32_t fd, void* buf, int32_t nbytes);
 extern int32_t sys_write(int32_t fd, const void* buf, int32_t nbytes);
